[TCA] bug fix (#2611)
* fix

* another one

* dexter fix

+++ parlai/core/torch_classifier_agent.py
@@ -282,7 +282,7 @@ class TorchClassifierAgent(TorchAgent):
         else:
             ref_prob = probs.cpu()[:, 0]
             # choose ref class if Prob(ref class) > threshold
-            prediction_id = ref_prob <= self.threshold
+            prediction_id = (ref_prob <= self.threshold).to(torch.int64)
         preds = [self.class_list[idx] for idx in prediction_id]
 
         if batch.labels is None or self.opt['ignore_labels']:
